client_templates/PLMB_TEMPLATE/requirements.txt
# Core dependencies - FastAPI version
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
twilio==8.10.0
deepgram-sdk==2.12.0
openai==1.3.0
elevenlabs==0.2.26
python-dotenv==1.0.0
jinja2==3.1.2


# Calendar integration dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytz==2023.3

# Additional utilities
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2

# Session and cache dependencies
redis==4.5.4